+++ zc-cache-spring-boot-starter/src/main/java/com/zhengcheng/cache/CaffeineCacheAutoConfiguration.java
@@ -1,4 +1,4 @@
-package com.zhengcheng.redis;
+package com.zhengcheng.cache;
 
 import com.github.benmanes.caffeine.cache.Cache;
 import com.github.benmanes.caffeine.cache.Caffeine;