+++ zc-cache-spring-boot-starter/src/main/java/com/zhengcheng/cache/RedisAutoConfiguration.java
@@ -1,4 +1,4 @@
-package com.zhengcheng.redis;
+package com.zhengcheng.cache;
 
 import com.fasterxml.jackson.annotation.JsonAutoDetect;
 import com.fasterxml.jackson.annotation.PropertyAccessor;