+++ zc-cache-spring-boot-starter/src/main/java/com/zhengcheng/cache/RequestLimitConfiguration.java
@@ -1,7 +1,7 @@
-package com.zhengcheng.redis;
+package com.zhengcheng.cache;
 
 import cn.hutool.core.util.StrUtil;
-import com.zhengcheng.redis.annotation.RequestLimit;
+import com.zhengcheng.cache.annotation.RequestLimit;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.BeansException;
 import org.springframework.context.ApplicationContext;